Add log_plays.py plugin
This plugins allows logging run results to file, in addition to (and in
a more verbose format than) the default logging. It is normally
available from the community plugins bundle, but to allow making some
improvements to it, the plugin is added here (where it will be found by
ansible automatically).

This file is taken verbatim from:

  https://github.com/ansible-collections/community.general/blob/9946f758af5fe0fe41f1cf7584d670ca1d6c2d52/plugins/callback/log_plays.py