Add changed_when to all command invocations
For builtin modules, ansible already knows if the task actually made
a change, but for commands it can only assume that changes were made.
using the `when_changed` key, ansible can be told exactly what to check
for (in stdout in this case).

This makes sure that in the command output, a task only shows as
"changed" when it actually changed something, and prevents running
handlers when they are not needed.